requirements.txt
dagster==1.9.0
dagster-webserver==1.9.0
PyYAML==6.0.2
httpx==0.28.0
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pytest==8.3.0
